refactor: improve project structure with core, shared and feature-based organization
