+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, signal } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { FooterComponent } from './shared/components/footer/footer.component';
+import { FooterComponent } from './shared/layout/footer/footer.component';
 import { AuthService } from './core/auth/auth.service';
 import { HeaderComponent } from './shared/layout/header/header.component';
 import { CommonModule } from '@angular/common';
@@ -9,8 +9,8 @@ import { CommonModule } from '@angular/common';
   selector: 'app-root',
   standalone: true,
   imports: [FooterComponent, RouterOutlet, HeaderComponent, CommonModule],
-  templateUrl: './app.html',
-  styleUrls: ['./app.scss'],
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss'],
 })
 export class App {
