+++ src/app/app.config.ts
@@ -6,8 +6,8 @@ import { MatPaginatorIntl } from '@angular/material/paginator';
 
 import { routes } from './app.routes';
 import { provideHttpClient, withInterceptors } from '@angular/common/http';
-import { authInterceptor } from './core/api/auth.interceptor';
-import { getPortuguesePaginatorIntl } from './shared/paginator-itl';
+import { authInterceptor } from './core/http/credentials.interceptor';
+import { getPortuguesePaginatorIntl } from './shared/paginator-intl';
 
 
 export const appConfig: ApplicationConfig = {
