+++ src/app/app.routes.ts
@@ -1,10 +1,10 @@
 import { Routes } from '@angular/router';
-import { authGuard } from './core/auth/auth-guard';
+import { authGuard } from './core/auth/auth.guard';
 import { LoginComponent } from './features/auth/login/login.component';
 import { HomeComponent } from './features/home/home.component';
 import { PersonComponent } from './features/person/pages/list/person-list.component';
 import { PersonCreateComponent } from './features/person/pages/create/person-create.component';
-import { PersonEditComponent } from './features/person/pages/person-edit/person-edit.component';
+import { PersonEditComponent } from './features/person/pages/edit/person-edit.component';
 
 export const routes: Routes = [
