+++ src/app/shared/components/error-snackbar/error-snackbar.component.scss
