+++ src/app/shared/components/filter-panel/filter-panel.component.scss
