+++ src/app/shared/layout/footer/footer.component.scss
