+++ src/app/shared/layout/header/header.component.scss
