+++ tsconfig.spec.json
@@ -11,5 +11,5 @@
   "include": [
     "src/**/*.d.ts",
     "src/**/*.spec.ts"
-  ]
+, "src/app/shared/components/button/button.component.spec.ts"  ]
 }
